Refactor package auth
