+++ app/src/main/java/app/tmdb/test/ui/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package app.tmdb.test.ui.auth
+package app.tmdb.test.ui.login
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
